update the install.sh to support the airgap installation

+++ install.sh
@@ -1,4 +1,10 @@
 #!/bin/sh
-npm install --unsafe-perm
+
+if [ -f ./node_modules.tgz ]
+then
+    tar -xvf ./node_modules.tgz
+else
+    npm install --unsafe-perm
+fi
 
 npm run build
